+++ examples/svelte/large-form/src/runes/form.ts
@@ -2,7 +2,7 @@ import { createFormCreator } from '@tanstack/svelte-form'
 import TextField from '../components/text-field.svelte'
 import SubscribeButton from '../components/subscribe-button.svelte'
 
-export const { createAppForm } = createFormCreator({
+export const { createAppForm, getFormType } = createFormCreator({
   fieldComponents: {
     TextField,
   },