+++ packages/form-core/src/FormApi.ts
@@ -49,7 +49,12 @@ import type {
   ValidationErrorMap,
   ValidationErrorMapKeys,
 } from './types'
-import type { DeepKeys, DeepKeysOfType, DeepValue } from './util-types'
+import type {
+  DeepKeys,
+  DeepKeysOfType,
+  DeepValue,
+  RejectPromiseValidator,
+} from './util-types'
 import type { Updater } from './utils'
 
 /**
@@ -188,11 +193,11 @@ export interface FormValidators<
   /**
    * Optional function that fires as soon as the component mounts.
    */
-  onMount?: TOnMount
+  onMount?: RejectPromiseValidator<TOnMount>
   /**
    * Optional function that checks the validity of your data whenever a value changes
    */
-  onChange?: TOnChange
+  onChange?: RejectPromiseValidator<TOnChange>
   /**
    * Optional onChange asynchronous counterpart to onChange. Useful for more complex validation logic that might involve server requests.
    */
@@ -204,7 +209,7 @@ export interface FormValidators<
   /**
    * Optional function that validates the form data when a field loses focus, returns a `FormValidationError`
    */
-  onBlur?: TOnBlur
+  onBlur?: RejectPromiseValidator<TOnBlur>
   /**
    * Optional onBlur asynchronous validation method for when a field loses focus returns a ` FormValidationError` or a promise of `Promise<FormValidationError>`
    */
@@ -213,9 +218,9 @@ export interface FormValidators<
    * The default time in milliseconds that if set to a number larger than 0, will debounce the async validation event by this length of time in milliseconds.
    */
   onBlurAsyncDebounceMs?: number
-  onSubmit?: TOnSubmit
+  onSubmit?: RejectPromiseValidator<TOnSubmit>
   onSubmitAsync?: TOnSubmitAsync
-  onDynamic?: TOnDynamic
+  onDynamic?: RejectPromiseValidator<TOnDynamic>
   onDynamicAsync?: TOnDynamicAsync
   onDynamicAsyncDebounceMs?: number
 }
@@ -662,16 +667,18 @@ export type DerivedFormState<
    * The error array for the form itself.
    */
   errors: Array<
-    | UnwrapFormValidateOrFn<TOnMount>
-    | UnwrapFormValidateOrFn<TOnChange>
-    | UnwrapFormAsyncValidateOrFn<TOnChangeAsync>
-    | UnwrapFormValidateOrFn<TOnBlur>
-    | UnwrapFormAsyncValidateOrFn<TOnBlurAsync>
-    | UnwrapFormValidateOrFn<TOnSubmit>
-    | UnwrapFormAsyncValidateOrFn<TOnSubmitAsync>
-    | UnwrapFormValidateOrFn<TOnDynamic>
-    | UnwrapFormAsyncValidateOrFn<TOnDynamicAsync>
-    | UnwrapFormAsyncValidateOrFn<TOnServer>
+    NonNullable<
+      | UnwrapFormValidateOrFn<TOnMount>
+      | UnwrapFormValidateOrFn<TOnChange>
+      | UnwrapFormAsyncValidateOrFn<TOnChangeAsync>
+      | UnwrapFormValidateOrFn<TOnBlur>
+      | UnwrapFormAsyncValidateOrFn<TOnBlurAsync>
+      | UnwrapFormValidateOrFn<TOnSubmit>
+      | UnwrapFormAsyncValidateOrFn<TOnSubmitAsync>
+      | UnwrapFormValidateOrFn<TOnDynamic>
+      | UnwrapFormAsyncValidateOrFn<TOnDynamicAsync>
+      | UnwrapFormAsyncValidateOrFn<TOnServer>
+    >
   >
   /**
    * A boolean indicating if any of the form fields are currently validating.
@@ -1041,6 +1048,7 @@ export class FormApi<
             isValidating: false,
             isBlurred: false,
             isDirty: false,
+            _arrayVersion: 0,
             ...(existingFieldMeta ?? {}),
             errorSourceMap: {
               ...(existingFieldMeta?.['errorSourceMap'] ?? {}),
@@ -1130,16 +1138,12 @@ export class FormApi<
           // As primitives, we don't need to aggressively persist the same referential value for performance reasons
           const isFieldValid = !isNonEmptyArray(fieldErrors)
           const isFieldPristine = !currBaseMeta.isDirty
-          const isDefaultValue =
-            evaluate(
-              curFieldVal,
+          const isDefaultValue = evaluate(
+            curFieldVal,
+            // eslint-disable-next-line @typescript-eslint/no-unnecessary-condition
+            this.getFieldInfo(fieldName)?.instance?.options.defaultValue ??
               getBy(this.options.defaultValues, fieldName),
-            ) ||
-            evaluate(
-              curFieldVal,
-              // eslint-disable-next-line @typescript-eslint/no-unnecessary-condition
-              this.getFieldInfo(fieldName)?.instance?.options.defaultValue,
-            )
+          )
 
           if (
             prevFieldInfo &&
@@ -1253,14 +1257,16 @@ export class FormApi<
       ) {
         errors = Object.values(currBaseStore.errorMap).reduce<
           Array<
-            | UnwrapFormValidateOrFn<TOnMount>
-            | UnwrapFormValidateOrFn<TOnChange>
-            | UnwrapFormAsyncValidateOrFn<TOnChangeAsync>
-            | UnwrapFormValidateOrFn<TOnBlur>
-            | UnwrapFormAsyncValidateOrFn<TOnBlurAsync>
-            | UnwrapFormValidateOrFn<TOnSubmit>
-            | UnwrapFormAsyncValidateOrFn<TOnSubmitAsync>
-            | UnwrapFormAsyncValidateOrFn<TOnServer>
+            NonNullable<
+              | UnwrapFormValidateOrFn<TOnMount>
+              | UnwrapFormValidateOrFn<TOnChange>
+              | UnwrapFormAsyncValidateOrFn<TOnChangeAsync>
+              | UnwrapFormValidateOrFn<TOnBlur>
+              | UnwrapFormAsyncValidateOrFn<TOnBlurAsync>
+              | UnwrapFormValidateOrFn<TOnSubmit>
+              | UnwrapFormAsyncValidateOrFn<TOnSubmitAsync>
+              | UnwrapFormAsyncValidateOrFn<TOnServer>
+            >
           >
         >((prev, curr) => {
           if (curr === undefined) return prev
@@ -1531,16 +1537,35 @@ export class FormApi<
       }
     }
 
-    this.baseStore.setState(() =>
-      getDefaultFormState({
+    this.baseStore.setState(() => {
+      let nextValues =
+        values ??
+        this.options.defaultValues ??
+        this.options.defaultState?.values
+
+      if (!values) {
+        ;(Object.values(this.fieldInfo) as FieldInfo<any>[]).forEach(
+          (fieldInfo) => {
+            if (
+              fieldInfo.instance &&
+              fieldInfo.instance.options.defaultValue !== undefined
+            ) {
+              nextValues = setBy(
+                nextValues,
+                fieldInfo.instance.name,
+                fieldInfo.instance.options.defaultValue,
+              )
+            }
+          },
+        )
+      }
+
+      return getDefaultFormState({
         ...(this.options.defaultState as any),
-        values:
-          values ??
-          this.options.defaultValues ??
-          this.options.defaultState?.values,
+        values: nextValues,
         fieldMetaBase,
-      }),
-    )
+      })
+    })
   }
 
   /**
@@ -2372,6 +2397,8 @@ export class FormApi<
       (prev) => [...(Array.isArray(prev) ? prev : []), value] as any,
       options,
     )
+
+    metaHelper(this).bumpArrayVersion(field)
   }
 
   insertFieldValue = async <TField extends DeepKeysOfType<TFormData, any[]>>(
@@ -2428,6 +2455,8 @@ export class FormApi<
       },
       mergeOpts(options, { dontValidate: true }),
     )
+
+    metaHelper(this).bumpArrayVersion(field)
 
     const dontValidate = options?.dontValidate ?? false
     if (!dontValidate) {
@@ -2560,6 +2589,8 @@ export class FormApi<
       mergeOpts(options, { dontValidate: true }),
     )
 
+    metaHelper(this).bumpArrayVersion(field)
+
     if (lastIndex !== null) {
       for (let i = 0; i <= lastIndex; i++) {
         const fieldKey = `${field}[${i}]`
@@ -2579,15 +2610,21 @@ export class FormApi<
    */
   resetField = <TField extends DeepKeys<TFormData>>(field: TField) => {
     this.baseStore.setState((prev) => {
+      const fieldDefault =
+        this.getFieldInfo(field).instance?.options.defaultValue
+      const formDefault = getBy(this.options.defaultValues, field)
+      const targetValue = fieldDefault ?? formDefault
+
       return {
         ...prev,
         fieldMetaBase: {
           ...prev.fieldMetaBase,
           [field]: defaultFieldMeta,
         },
-        values: this.options.defaultValues
-          ? setBy(prev.values, field, getBy(this.options.defaultValues, field))
-          : prev.values,
+        values:
+          targetValue !== undefined
+            ? setBy(prev.values, field, targetValue)
+            : prev.values,
       }
     })
   }
@@ -2662,16 +2699,18 @@ export class FormApi<
   getAllErrors = (): {
     form: {
       errors: Array<
-        | UnwrapFormValidateOrFn<TOnMount>
-        | UnwrapFormValidateOrFn<TOnChange>
-        | UnwrapFormAsyncValidateOrFn<TOnChangeAsync>
-        | UnwrapFormValidateOrFn<TOnBlur>
-        | UnwrapFormAsyncValidateOrFn<TOnBlurAsync>
-        | UnwrapFormValidateOrFn<TOnSubmit>
-        | UnwrapFormAsyncValidateOrFn<TOnSubmitAsync>
-        | UnwrapFormValidateOrFn<TOnDynamic>
-        | UnwrapFormAsyncValidateOrFn<TOnDynamicAsync>
-        | UnwrapFormAsyncValidateOrFn<TOnServer>
+        NonNullable<
+          | UnwrapFormValidateOrFn<TOnMount>
+          | UnwrapFormValidateOrFn<TOnChange>
+          | UnwrapFormAsyncValidateOrFn<TOnChangeAsync>
+          | UnwrapFormValidateOrFn<TOnBlur>
+          | UnwrapFormAsyncValidateOrFn<TOnBlurAsync>
+          | UnwrapFormValidateOrFn<TOnSubmit>
+          | UnwrapFormAsyncValidateOrFn<TOnSubmitAsync>
+          | UnwrapFormValidateOrFn<TOnDynamic>
+          | UnwrapFormAsyncValidateOrFn<TOnDynamicAsync>
+          | UnwrapFormAsyncValidateOrFn<TOnServer>
+        >
       >
       errorMap: ValidationErrorMap<
         UnwrapFormValidateOrFn<TOnMount>,